backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
python-dotenv
pydantic>=2.7
python-jose[cryptography]
passlib[bcrypt]
httpx
aiohttp>=3.9
pydantic-settings>=2.2
psycopg[binary]>=3.1
sqlalchemy>=2.0
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.2.1
click>=8.1.7
typer>=0.12.0
